+++ ui/src/components/ListPagingArticles/articleItem.tsx
@@ -1,13 +1,31 @@
 import React from "react"
+import { Mutation } from "react-apollo";
 
-import { Article } from "./container"
+import {DeleteArticle, DeleteArticleVariables} from "../__generated__/DeleteArticle"
+import { Article, PageInfo, deleteArticle, deleteUpdateArticle } from "./container"
 
 interface ArticleItemProps {
+    diaryId: string;
+    pageInfo: PageInfo;
+    article: Article;
+}
+
+export const ArticleItem: React.StatelessComponent<ArticleItemProps> = ({diaryId, pageInfo, article}) => (
+    <Mutation<DeleteArticle, DeleteArticleVariables> mutation={deleteArticle}>
+    {(deleteArticle) => (
+        <ArticleItemPresentation
+            article={article} 
+            deleteArticle={(articleId: string) => deleteArticle({ variables: {articleId}, update: deleteUpdateArticle(diaryId, articleId, pageInfo.currentPage) })}
+        />
+    )}
+    </Mutation>
+)
+
+interface ArticleItemPresentationProps {
     article: Article
     deleteArticle?: (articleId: string) => void
 }
-
-export const ArticleItem: React.StatelessComponent<ArticleItemProps> = ({article, deleteArticle}) => {
+const ArticleItemPresentation: React.StatelessComponent<ArticleItemPresentationProps> = ({article, deleteArticle}) => {
     return (
         <div className="DiaryArticle">
             <h2>{article.title}</h2>
