+++ ui/src/components/ListPagingArticles/container.tsx
@@ -1,9 +1,12 @@
 import React from "react"
 import gql from "graphql-tag"
+import { MutationUpdaterFn } from "apollo-client";
 
 import {AritlceListWithPagination} from "./articleListWithPagination"
+import {DeleteArticle} from "../__generated__/DeleteArticle"
+import { ListArticles} from "./__generated__/ListArticles";
 
-const listArticlesFragment = gql`
+export const listArticlesFragment = gql`
     fragment ListArticlesFragment on Article {
         id
         diaryId
@@ -34,6 +37,27 @@ export const listArticleQuery = gql`
     ${pageInfoFragment}
     ${listArticlesFragment}
 `
+
+export const deleteArticle = gql`
+    mutation DeleteArticle($articleId: ID!) {
+        deleteArticle(articleId: $articleId)
+    }
+`
+
+export const deleteUpdateArticle: (diaryId: string, articleId: string, page: number) => MutationUpdaterFn<DeleteArticle> = (diaryId, articleId, page) => (cache, result) => {
+    const { data } = result;
+    const listArticles = cache.readQuery<ListArticles>({ query: listArticleQuery, variables: {diaryId: diaryId, page: page}})       
+    if (listArticles && data) {
+        const articles = [...listArticles.listArticles.articles].filter(article => article.id !== articleId);
+        const newDiary = {
+            listArticles: {
+                ...listArticles.listArticles,
+                articles: articles,
+            }
+        };
+        cache.writeQuery({ query: listArticleQuery, variables: {diaryId: diaryId, page: page} , data: newDiary });
+    };
+}
 
 export type Article = {
     id: string,
