+++ ui/src/components/app.tsx
@@ -8,11 +8,11 @@ import {HttpLink} from "apollo-link-http";
 import {onError} from "apollo-link-error";
 import {InMemoryCache} from "apollo-cache-inmemory";
 
-import {GlobalHeader} from "./global_header";
-import {Index} from "./index";
-import {Diary} from "./diary";
-import {AddArticle} from "./addArticle"
-import {Me} from "./me"
+import {GlobalHeader} from "./GrobalHeader/global_header";
+import {UserTopContainer} from "./UserTop/container";
+import {DiaryTopContainer} from "./DiaryTop/container";
+import {AddArticleFormContainer} from "./AddArticleForm/container"
+import {Me} from "./Me/me"
 
 const client = new ApolloClient({
   link: ApolloLink.from([
@@ -40,10 +40,10 @@ export const App: React.StatelessComponent = () => (
         <GlobalHeader />
         <main>
           <Switch>
-            <Route exact path="/" component={Index} />
+            <Route exact path="/" component={UserTopContainer} />
             <Route exact path="/me" component={Me} />
-            <Route exact path="/diaries/:diaryId/add" component={AddArticle} />
-            <Route exact path="/diaries/:diaryId/:page?" component={Diary} />
+            <Route exact path="/diaries/:diaryId/add" component={AddArticleFormContainer} />
+            <Route exact path="/diaries/:diaryId/:page?" component={DiaryTopContainer} />
           </Switch>
         </main>
       </>
